Change logging of spectrogram manipulation options
When a manipulation function is called, we want to log what
keyword arguments it was called with, including the default
options for any keyword arguments that were not provided.

Previously, this was done by calling locals() at the top of
every manipulation function, then manually deleting the
'spectrogram' value in the dictionary of locals. These options
would be returned at the bottom of the manipulation function,
and the _spectrogram_manipulation wrapper would then add the
manipulation keywords to the Spectrogram instance.

Now, this method is simplified. All manipulation functions
accept spectrogram only as a positional argument, not as
a keyword argument. The manipulation functions also only
return the filtered spectrogram to the wrapper function,
whereas before the wrapper function received both the filtered
spectrogram and the manipulation options.

Within the wrapper, a dictionary of options is created by
combining any keyword arguments that the function was called
with, along with any defaults that the function used. The wrapper
then uses this dictionary to log the spectrogram manipulation.